Further reorganized item runtime
