+++ src/sheets/item/ClassSheet.svelte
@@ -8,7 +8,7 @@
   import { FoundryAdapter } from 'src/foundry/foundry-adapter';
   import ItemProfilePicture from './parts/ItemProfilePicture.svelte';
   import TextInput from 'src/components/inputs/TextInput.svelte';
-  import itemSheetTabs from '../../runtime/itemSheetTabs';
+  import itemSheetTabs from '../../runtime/item/itemSheetTabs';
   import Source from '../shared/Source.svelte';
 
   let context = getContext<Readable<ItemSheetContext>>('context');
